# Improve mavlink compatility performance