This patch improves by 96% the translation from Packet into rust-mavlink raw messages.

![image](https://github.com/user-attachments/assets/4c20861e-f708-4aac-9bf7-fd309da2b1dd)


## New:

<HTML>
<table>
    <thead>
        <tr>
            <th></th>
            <th title="0.95 confidence level">Lower bound</th>
            <th>Estimate</th>
            <th title="0.95 confidence level">Upper bound</th>
        </tr>
    </thead>
    <tbody>
        <tr>
            <td>Slope</td>
            <td>726.70 µs</td>
            <td>726.82 µs</td>
            <td>726.95 µs</td>
        </tr>
        <tr>
            <td>Throughput</td>
            <td>13.756 Melem/s</td>
            <td>13.759 Melem/s</td>
            <td>13.761 Melem/s</td>
        </tr>
        <tr>
            <td>R²</td>
            <td>0.9999140</td>
            <td>0.9999179</td>
            <td>0.9999132</td>
        </tr>
        <tr>
            <td>Mean</td>
            <td>726.81 µs</td>
            <td>727.06 µs</td>
            <td>727.41 µs</td>
        </tr>
        <tr>
            <td title="Standard Deviation">Std. Dev.</td>
            <td>625.11 ns</td>
            <td>1.5624 µs</td>
            <td>2.4654 µs</td>
        </tr>
        <tr>
            <td>Median</td>
            <td>726.57 µs</td>
            <td>726.64 µs</td>
            <td>726.80 µs</td>
        </tr>
        <tr>
            <td title="Median Absolute Deviation">MAD</td>
            <td>271.57 ns</td>
            <td>391.99 ns</td>
            <td>571.30 ns</td>
        </tr>
    </tbody>
</table>
</HTML>

## Old:

<HTML>
<table>
    <thead>
        <tr>
            <th></th>
            <th title="0.95 confidence level">Lower bound</th>
            <th>Estimate</th>
            <th title="0.95 confidence level">Upper bound</th>
        </tr>
    </thead>
    <tbody>
        <tr>
            <td>Slope</td>
            <td>1.4197 ms</td>
            <td>1.4202 ms</td>
            <td>1.4207 ms</td>
        </tr>
        <tr>
            <td>Throughput</td>
            <td>7.0389 Melem/s</td>
            <td>7.0414 Melem/s</td>
            <td>7.0436 Melem/s</td>
        </tr>
        <tr>
            <td>R²</td>
            <td>0.9993518</td>
            <td>0.9993660</td>
            <td>0.9993493</td>
        </tr>
        <tr>
            <td>Mean</td>
            <td>1.4235 ms</td>
            <td>1.4257 ms</td>
            <td>1.4282 ms</td>
        </tr>
        <tr>
            <td title="Standard Deviation">Std. Dev.</td>
            <td>8.1967 µs</td>
            <td>12.247 µs</td>
            <td>15.793 µs</td>
        </tr>
        <tr>
            <td>Median</td>
            <td>1.4200 ms</td>
            <td>1.4206 ms</td>
            <td>1.4216 ms</td>
        </tr>
        <tr>
            <td title="Median Absolute Deviation">MAD</td>
            <td>2.2494 µs</td>
            <td>3.0565 µs</td>
            <td>4.7240 µs</td>
        </tr>
    </tbody>
</table>
</HTML>